+++ app/utils/defaultSettings.js
@@ -81,6 +81,7 @@ export default {
   hideStrictModePreferences: false,
   miniBreakManualFinish: false,
   longBreakManualFinish: false,
+  breakHealthMode: false,
   openAtLogin: false,
   _migratedOpenAtLogin: false
 }